Revisions to the capability mechanism per review
